% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ameacadas_br}
\alias{ameacadas_br}
\title{Threatened Plant Species in Brazil (Portaria MMA nº 148/2022)}
\format{
A data frame. Column names and structure depend on the original CSV.
}
\source{
Ministério do Meio Ambiente (MMA), Portaria nº 148, de 7 de junho de 2022.
}
\usage{
ameacadas_br
}
\description{
A dataset containing the list of plant species classified as threatened in Brazil,
according to the Brazilian Ministry of the Environment's ordinance Portaria MMA nº 148, published in 2022.
Please check the original document after use this list (https://www.in.gov.br/en/web/dou/-/portaria-mma-n-148-de-7-de-junho-de-2022-406272733).
}
\seealso{
\code{\link{ameacadas_rs}}
}
\keyword{datasets}
