% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ameacadas_br_cnc2020}
\alias{ameacadas_br_cnc2020}
\title{Threatened Brazilian Plants — CNCFlora 2020}
\format{
A data frame. Column names and structure depend on the original CSV.
}
\source{
CNCFlora (Centro Nacional de Conservação da Flora), 2020.
}
\usage{
ameacadas_br_cnc2020
}
\description{
A dataset listing threatened plant species in Brazil according to the CNCFlora (Centro Nacional de Conservação da Flora) 2020 assessment.
Please check the original source after use this list (https://cncflora.jbrj.gov.br/).
}
\seealso{
\code{\link{ameacadas_br}}
}
\keyword{datasets}
