% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ameacadas_rs}
\alias{ameacadas_rs}
\title{Threatened Plant Species in Rio Grande do Sul (Decreto Estadual nº 52.109/2014)}
\format{
A data frame. Column names and structure depend on the original CSV.
}
\source{
Governo do Estado do Rio Grande do Sul, Decreto nº 52.109, de 1º de dezembro de 2014.
}
\usage{
ameacadas_rs
}
\description{
A dataset listing plant species classified as threatened in the state of Rio Grande do Sul, Brazil,
based on the state decree Decreto nº 52.109, published in 2014.
Please check the original source after use this list (https://www.al.rs.gov.br/legis/M010/M0100099.ASP?Hid_Tipo=TEXTO&Hid_TodasNormas=61669&hTexto=&Hid_IDNorma=61669).
}
\seealso{
\code{\link{ameacadas_br}}
}
\keyword{datasets}
