% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPeq.R
\name{DAPeq}
\alias{DAPeq}
\title{DAPeq}
\usage{
DAPeq(x, ff)
}
\arguments{
\item{x}{A data frame containing the columns: plots (\code{parc}), species (\code{spp}), height (\code{H}), and either diameters (\code{dap}) or circumferences (\code{cap}).}

\item{ff}{The form factor to use in calculations. Defaults to \code{0.5} if not provided.}
}
\value{
A new data frame with the equivalent diameter and plant volumes, calculated based on the specified form factor.
}
\description{
Function DAPeq
}
\details{
Calculates the equivalent diameter for multi-stemmed plants.

The equivalent diameter is calculated for plants with multi-stemmed trunks. The results include the volume of each plant, considering the provided or default form factor.
}
\examples{
data(Southern_Forest)
dap.eq(x = Southern_Forest, ff = 0.5)
}
