% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReflora.R
\name{getReflora}
\alias{getReflora}
\title{Download and Import Taxonomic Data from Flora and Funga of Brazil}
\usage{
getReflora(x, dir = tempdir(), simplify = TRUE, cores = 3)
}
\arguments{
\item{dir}{Character. The directory where the DwC-A file will be downloaded and extracted.
Default is a temporary directory created with \code{tempdir()}.}

\item{simply}{Logical. If \code{TRUE}, removes some columns related to taxonomic hierarchy and remarks
(e.g., phylum, class, order, occurrenceRemarks) to simplify the output. Default is \code{TRUE}.}
}
\value{
A \code{data.frame} containing taxonomic information from Flora and Funga of Brazil. Columns include:
\itemize{
\item \code{id} - Unique identifier for each taxon
\item \code{family}, \code{genus}, \code{taxa}, \code{scientificName}, \code{acceptedNameUsage}
\item \code{taxonomicStatus}, \code{nomenclaturalStatus}, \code{Endemismo}, \code{lifeForm}, and more
}
The exact columns may vary depending on the \code{simply} argument.
}
\description{
This function downloads the latest version of the Darwin Core Archive (DwC-A) from the official
Flora and Funga of Brazil repository, extracts and processes taxonomic, distribution, and species
profile information. The result is returned as a cleaned and optionally simplified \code{data.frame}.
}
\details{
The function processes three files from the Darwin Core Archive:
\itemize{
\item \code{taxon.txt}: Main taxonomic data
\item \code{distribution.txt}: Information about geographic distribution and endemism
\item \code{speciesprofile.txt}: Life form and other traits
}
It also filters for Plantae and constructs species names using taxonomic rank information.
}
\note{
This function requires internet access and may take several seconds to run.
}
\examples{
\dontrun{
flora <- getReflora(cores = 5)
head(flora)
}

}
