% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getThreat.R
\name{getSpp}
\alias{getSpp}
\title{Matching species names with a species df}
\usage{
getSpp(species, species_to_match, max.distance = 0.05, ncores = 1)
}
\arguments{
\item{species}{A character vector with the species names to be matched.}

\item{species_to_match}{A data frame of threaten species with at least one columns: \code{spp} (reference species names).}

\item{max.distance}{A numeric value between 0 and 1 indicating the maximum allowed distance for matching (default: 0.05).}

\item{ncores}{The number of cores to use for parallel processing (default: 1).}
}
\value{
A data frame containing, for each matched input species:
\itemize{
\item \code{input_name}: the original species name provided.
\item \code{Suggested_name}: the best-matched species name from the reference list.
\item \code{Distance}: the string edit distance between the input and suggested name.
\item \code{Categoria}: the associated category of the matched species.
}
}
\description{
Performs approximate (fuzzy) matching between a list of species names and a reference list, using \code{agrep} with parallel processing.
The main goal of this function is to identify threatened species in your dataset based on the reference data provided within the package (\code{ameacadas_rs}, \code{ameacadas_br}, \code{ameacadas_br_cnc2020}, \code{plantas_raras_br}).
}
\details{
This function uses \code{agrep} internally to perform fuzzy matching and leverages \code{doParallel} for multi-core processing. Only input species with at least one match are included in the result.
}
\examples{
\dontrun{
data(ameacadas_rs)
species <- c("Araucaria angustifolia", "Hippeastrum breviflorum")
species_to_match <- ameacadas_rs
getThreat(species, species_to_match, max.distance = 0.1, ncores = 2)
}

}
