% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herbPhyto.R
\name{herbPhyto}
\alias{herbPhyto}
\title{plant_cover}
\usage{
herbPhyto(x, area)
}
\arguments{
\item{x}{A data frame containing the columns: plots (\code{parc}), species (\code{spp}), and absolute cover (\code{cob}).}

\item{area}{The area of the sampling units in square meters.}
}
\value{
A list summarizing the data, including the number of sampling units, total sampled area, species richness, Shannon-Wiener diversity index (in nats, based on cover), and Pielou's evenness index.
}
\description{
The \code{plantCover} function calculates phytosociological parameters for herbaceous vegetation based on cover values.
}
\details{
The input file must include the following columns with exact lowercase names:
\itemize{
\item \code{parc}: plot identification (numeric or character, either works)
\item \code{spp}: common or scientific name of the individual
\item \code{cob}: absolute cover percentage
}
}
\examples{
Example usage:
data(Southern_Camp)
herbPhyto(x = Southern_Camp, area = 1)
}
