% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plantas_raras_br}
\alias{plantas_raras_br}
\title{Rare Plant Species in Brazil}
\format{
A data frame. Column names and structure depend on the original CSV.
}
\source{
Internal compilation from: Guillet et al. (2009) Plantas Raras do Brasil
}
\usage{
plantas_raras_br
}
\description{
A dataset listing rare native plant species in Brazil. The criteria for rarity may include geographic distribution, endemism, and population size.
}
\keyword{datasets}
