% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPhyto.R
\name{reportphyto}
\alias{reportphyto}
\title{Generate a Phytosociological Report}
\usage{
reportphyto(
  x,
  type = "forest",
  salvar_txt = FALSE,
  arquivo_txt = "relatorio_fitossociologico.txt",
  lang = "pt"
)
}
\arguments{
\item{x}{A list containing two components:
\itemize{
\item \code{Phytossociology}: A data frame with phytosociological metrics. For \code{type = "forest"}, the columns must include
\code{N} (density), \code{AB} (basal area), \code{FA} (frequency), \code{DoA} (dominance), and \code{VI} (importance value).
For \code{type = "herb"}, the columns must include \code{FA} (frequency), \code{CA} (coverage), and \code{VI} (importance value).
\item \code{Resume}: A data frame with summarized metrics (e.g., total individuals, species richness, Shannon-Wiener diversity, etc.).
}}

\item{type}{Character. Type of vegetation. Use \code{"forest"} for tree data or \code{"herb"} for herbaceous data. Default is \code{"forest"}.}

\item{salvar_txt}{Logical. If \code{TRUE}, the generated report will be saved as a \code{.txt} file. Default is \code{FALSE}.}

\item{arquivo_txt}{Character. Name of the \code{.txt} file where the report will be saved. Default is \code{"relatorio_fitossociologico.txt"}.}

\item{lang}{Character. Language of the report. Use \code{"pt"} for Portuguese or \code{"us"} for English. Default is \code{"pt"}.}
}
\value{
A character string containing the phytosociological report.
If \code{salvar_txt = TRUE}, the report is also saved as a \code{.txt} file.
}
\description{
This function generates a descriptive phytosociological report based on forest or herbaceous data.
It calculates metrics such as density, basal area, Shannon-Wiener diversity, and species importance values.
The report can be generated in Portuguese or English and optionally saved as a \code{.txt} file.

WE ADVISE THAT THIS FUNCTION DON'T SUBSTITUTE A GOOD LOOK INTO YOUR DATA. WE ALSO ADVISE AGAINST THE RESULTS COPY AND PASTE.
}
\examples{
data("Southern_Forest")
data("Southern_Camp")

data <- phytos::PhytoIndex(Southern_Forest, 100)
data2 <- phytos::PhytoIndex(Southern_Camp, 100)

# Generate a forest report in Portuguese
reportphyto(data, type = "forest", n = 3, salvar_txt = FALSE, lang = "pt")

# Generate a herbaceous report in English and save to file
reportphyto(data2, type = "herb", salvar_txt = TRUE, arquivo_txt = "herb_report.txt", lang = "us")

}
