% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveSpp.R
\name{resolveSpp}
\alias{resolveSpp}
\title{Resolve Species Names Using Flora and Funga of Brazil}
\usage{
resolveSpp(species, flora = getReflora(), simplify = FALSE)
}
\arguments{
\item{species}{Character vector. List of species names to be resolved. Names should follow the botanical
binomial convention (e.g., "Araucaria angustifolia"), morphospecies nomination is allowed (e.g., Cyperus sp. 1 or Asteraceae sp. 1)
but will be concatenated in the final data frame.}

\item{flora}{A \code{data.frame} returned by \code{\link[=getReflora]{getReflora()}}. If not provided, the function will call \code{getReflora()}
to download and process the most recent dataset. Default is \code{getReflora()}.}

\item{simplify}{Logical. If \code{TRUE} return few columns with species submitted, matched and acceptedname.
If \code{FALSE} return all previous information along with species origin, habitat, lifeform or its occurrence domain.}
}
\value{
A \code{data.frame} with taxonomic resolution results, including the following columns (if available):
\itemize{
\item \code{Family} — Family name
\item \code{Genus} — Genus name
\item \code{SubmittedName} — Original species name submitted by the user
\item \code{MatchedName} — Closest matching name found in the flora
\item \code{AcceptedName} — Accepted name based on taxonomic status (2 or 3-word version)
\item \code{MatchType} — Match type: "Exact", "Fuzzy", or "Not Found"
\item \code{Status} — Taxonomic status (e.g., "Accepted Name" or "Synonym")
\item \code{Endemism} — Endemism status, if available (e.g., "Endêmica", "Não Endêmica")
\item \code{Origin} — Return if species is native, cultivated or naturalized
\item \code{VegType} — Brazilian phytogeographic domain of occurrence
}
}
\description{
This function correct your list of plant species names based on Flora and Funga of Brazil \url{https://floradobrasil.jbrj.gov.br/consulta/#CondicaoTaxonCP}
using exact and fuzzy string matching. It returns a cleaned and enriched data frame with taxonomic
status, accepted names, endemism, and origin.
}
\details{
The function correct performs:
\itemize{
\item Exact matching
\item Fuzzy matching (with \code{max.distance = 0.1})
}
}
\note{
For large species lists, fuzzy matching may take a few seconds to process.

ALWAYS submit scientific names to minimize the errors.
ALWAYS check the final results looking if the submitted names matched correctly
}
\examples{
\dontrun{
# Load the Flora and Funga of Brazil dataset
flora <- getReflora()

# Example species list (Southern_Camp$spp must be available in your environment)
my_spp <- resolveSpp(Southern_Camp$spp, flora = getReflora())
head(my_spp)
}

}
\seealso{
\code{\link[=getReflora]{getReflora()}}
}
