% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Southern_Camp}
\alias{Southern_Camp}
\title{Herbaceous Vegetation Dataset}
\format{
A data frame with the following columns:
\describe{
\item{parc}{Character. Plot ID.}
\item{spp}{Character. Species name.}
\item{coverage}{Numeric. Coverage percentage.}
\item{height}{Numeric. Height (m).}
}
}
\usage{
Southern_Camp
}
\description{
A dataset with sample data for herbaceous vegetation, including plot ID, species, coverage, and height.
}
\examples{
data(Southern_Camp)
head(Southern_Camp)
}
\keyword{datasets}
