% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Southern_Forest}
\alias{Southern_Forest}
\title{Forest Vegetation Dataset}
\format{
A data frame with the following columns:
\describe{
\item{parc}{Character. Plot ID.}
\item{spp}{Character. Species name.}
\item{dap}{Numeric. Diameter at breast height (cm).}
\item{H}{Numeric. Height (m).}
}
}
\usage{
Southern_Forest
}
\description{
A dataset with sample data for forest vegetation, including plot ID, species, diameter at breast height, and height.
}
\examples{
data(Southern_Forest)
head(Southern_Forest)
}
\keyword{datasets}
