% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePhyto.R
\name{treePhyto}
\alias{treePhyto}
\title{treePhyto}
\usage{
treePhyto(x, area, filename = NULL, VI = "percent")
}
\arguments{
\item{x}{A data frame containing the columns: plots (\code{parc}), species (\code{spp}), height (\code{H}), and either diameters (\code{dap}) or circumferences (\code{cap}).}

\item{area}{The area of the sampling units in square meters.}

\item{filename}{The name of the output file for the results, provided as a string (single or double quotes), default is NULL.}

\item{VI}{The type of importance value. Options are \code{"cottam"} for the sum of relativized parameters, or \code{"percent"} (default) for the mean of relativized parameters.}
}
\value{
A list summarizing the data, including:
\itemize{
\item Basal area (m²/ha)
\item Density (ind./ha)
\item Richness (excluding dead specimens from the raw data)
\item Shannon-Wiener diversity index (in nats)
\item Pielou's evenness index
}
}
\description{
Function for calculating phytosociological descriptors and adjusting phytosociological data tables
}
\details{
This function calculates phytosociological descriptors and performs adjustments on data tables.
And is an adaptation of \code{FitoR} function available at https://github.com/ricds/fitoR, created by
Alexandre Gabriel Christo, Pedro Higuchi, Ricardo Dal'Agnol, and Arthur Vinicius Rodrigues.

The input file must contain the following columns with exact lowercase names:
\itemize{
\item \code{parc}: plot identification (numeric or character, either works)
\item \code{spp}: common or scientific name of the individual
\item \code{dap}: diameter at breast height (DBH) in cm
or
\item \code{cap}: circumference at breast height (CBH) in cm
}

It is advisable to use separate columns for individuals with multi-stemmed trunks, named as: \code{cap} or \code{dap1}, \code{dap2}, \code{dap3}, \code{dap4}, etc. Alternatively, use the equivalent diameter. Do NOT separate the stems of individuals into rows.

Note: The script will search for the keywords \code{cap} and \code{dap} in column names to locate them.
Therefore, avoid having extra columns containing these keywords in their names (e.g., a column named \code{dap_medio} could cause calculation errors).
}
\examples{
data(Southern_Forest)
treePhyto(x = Southern_Forest, area = 100, VI = "cottam", filename = "my_result")
}
